+++ src/components/icons/RemoveCart.tsx
@@ -1,4 +1,4 @@
-const RemoveCart = (props: React.SVGProps<SVGSVGElement>) => {
+export default function RemoveCart(props: React.SVGProps<SVGSVGElement>) {
   return (
     <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16" fill="none" {...props}>
       <g clipPath="url(#clip0_1275_13681)">
@@ -14,6 +14,4 @@ const RemoveCart = (props: React.SVGProps<SVGSVGElement>) => {
       </defs>
     </svg>
   );
-};
-
-export default RemoveCart;
+}
