+++ src/components/icons/TopArrow.tsx
@@ -1,4 +1,4 @@
-const TopArrow = (props: React.SVGProps<SVGSVGElement>) => {
+export default function TopArrow(props: React.SVGProps<SVGSVGElement>) {
   return (
     <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16" fill="none" {...props}>
       <path
@@ -10,6 +10,4 @@ const TopArrow = (props: React.SVGProps<SVGSVGElement>) => {
       />
     </svg>
   );
-};
-
-export default TopArrow;
+}
