+++ src/hooks/index.ts
@@ -1,2 +1,1 @@
-export { default as useCartItems } from "./useCartItems";
-export { default as useProducts } from "./useFetchProducts";
+export { default as useFetch } from "./useFetch";
